Crownix/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-docx==0.8.11
pdfplumber==0.9.0
pypdf==4.3.1
pytesseract==0.3.10
openai==1.3.0
Pillow==10.0.1
python-multipart==0.0.6
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7

